// src/app/Models/PsicologaRequest.ts
export interface PsicologaRequest{
    nome: string;
    dataNascimento: Date;
    rg: string;
    cpf: string;
    responsavel: string;
    endereco: string;
    municipio: string;
    saude: string;
    consciente: string;
    religiao: string;
    filho: string;
    qntFilho: string;
    acolhimento: string;
    historico: string;
}